apps/worker/requirements.txt
sentence-transformers
tqdm
pinecone
python-dotenv
fastapi
uvicorn[standard]

